// frontend/src/lib/api/model/subjectDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * e-proshno API
 * OpenAPI spec version: v1
 */

export interface SubjectDto {
  id: string;
  levelId: string;
  nameBn: string;
  /** @nullable */
  paper?: number | null;
  label: string;
  /** @nullable */
  code?: string | null;
  isCustom: boolean;
}